Split oversized eval solve modules
